plugins/aks-desktop: Extract ScalingCard logic into hooks
