+++ plugins/aks-desktop/src/components/Scaling/ScalingCard.tsx
@@ -2,28 +2,14 @@
 // Licensed under the Apache 2.0.
 
 import { Icon } from '@iconify/react';
-import { K8s } from '@kinvolk/headlamp-plugin/lib';
-import {
-  Box,
-  CircularProgress,
-  FormControl,
-  Grid,
-  InputLabel,
-  MenuItem,
-  Select,
-  Typography,
-} from '@mui/material';
-import React, { useCallback, useEffect, useState } from 'react';
-import {
-  CartesianGrid,
-  Legend,
-  Line,
-  LineChart,
-  ResponsiveContainer,
-  Tooltip,
-  XAxis,
-  YAxis,
-} from 'recharts';
+import { Box, Typography } from '@mui/material';
+import React from 'react';
+import { DeploymentSelector } from './components/DeploymentSelector';
+import { ScalingChart } from './components/ScalingChart';
+import { ScalingMetrics } from './components/ScalingMetrics';
+import { useChartData } from './hooks/useChartData';
+import { useDeployments } from './hooks/useDeployments';
+import { useHPAInfo } from './hooks/useHPAInfo';
 
 export interface ProjectDefinition {
   id: string;
@@ -37,239 +23,23 @@ interface ScalingCardProps {
   project: Project;
 }
 
-interface HPAInfo {
-  name: string;
-  namespace: string;
-  minReplicas: number;
-  maxReplicas: number;
-  targetCPUUtilization: number;
-  currentCPUUtilization: number;
-  currentReplicas: number;
-  desiredReplicas: number;
-}
-
-interface Deployment {
-  name: string;
-  namespace: string;
-  replicas: number;
-  availableReplicas: number;
-  readyReplicas: number;
-}
-
-interface ChartDataPoint {
-  time: string;
-  Replicas: number;
-  CPU: number;
-}
-
 function ScalingCard({ project }: ScalingCardProps) {
-  const [selectedDeployment, setSelectedDeployment] = useState<string>('');
-  const [deployments, setDeployments] = useState<Deployment[]>([]);
-  const [hpaInfo, setHpaInfo] = useState<HPAInfo | null>(null);
-  const [loading, setLoading] = useState<boolean>(false);
-  const [error, setError] = useState<string | null>(null);
+  const namespace = project.namespaces?.[0];
+  const cluster = project.clusters?.[0];
 
   // Fetch real deployments from Kubernetes API
-  const fetchDeployments = useCallback(async () => {
-    if (!project.namespaces?.[0]) return;
-
-    setLoading(true);
-    setError(null);
-
-    try {
-      // Use Headlamp's K8s API to fetch deployments
-      const cancel = K8s.ResourceClasses.Deployment.apiList(
-        (deploymentList: K8s.Deployment[]) => {
-          const deployments = deploymentList
-            .filter(
-              (deployment: K8s.Deployment) => deployment.getNamespace() === project.namespaces[0]
-            )
-            .map((deployment: K8s.Deployment) => ({
-              name: deployment.getName(),
-              namespace: deployment.getNamespace(),
-              replicas: deployment.spec?.replicas || 0,
-              availableReplicas: deployment.status?.availableReplicas || 0,
-              readyReplicas: deployment.status?.readyReplicas || 0,
-            }));
-
-          setDeployments(deployments);
-
-          // Auto-select first deployment if none selected
-          if (deployments.length > 0 && !selectedDeployment) {
-            setSelectedDeployment(deployments[0].name);
-          }
-          setLoading(false);
-        },
-        (error: any) => {
-          console.error('Error fetching deployments:', error);
-          setError('Failed to fetch deployments');
-          setDeployments([]);
-          setLoading(false);
-        },
-        {
-          namespace: project.namespaces[0],
-          cluster: project.clusters[0],
-        }
-      )();
-
-      // Return cleanup function
-      return cancel;
-    } catch (err) {
-      console.error('Error in fetchDeployments:', err);
-      setError('Failed to fetch deployments');
-      setLoading(false);
-    }
-  }, [project.namespaces, selectedDeployment]);
-
-  // Generate chart data based on the selected deployment and HPA info
-  const generateChartData = (): ChartDataPoint[] => {
-    const data: ChartDataPoint[] = [];
-    const now = new Date();
-
-    // Get current deployment info
-    const currentDeployment = deployments.find(d => d.name === selectedDeployment);
-
-    // Use actual data - no fake fallbacks
-    const currentReplicas = hpaInfo?.currentReplicas || currentDeployment?.readyReplicas || 0;
-    const currentCPU = hpaInfo?.currentCPUUtilization || 0; // Keep 0 if no real data
-
-    // Generate data for the last 24 hours (every 2 hours to avoid crowding)
-    for (let i = 23; i >= 0; i -= 2) {
-      const time = new Date(now.getTime() - i * 60 * 60 * 1000);
-      const timeString = `${time.getHours().toString().padStart(2, '0')}:00`;
-
-      let replicas = currentReplicas;
-      let cpu = currentCPU;
-
-      if (i === 0) {
-        // Current time - use actual values only
-        replicas = currentReplicas;
-        cpu = currentCPU;
-      } else {
-        // Historical data - only simulate if we have real current data
-        if (currentCPU > 0) {
-          // We have real CPU data, simulate historical variation
-          const timeVariation = Math.sin((i / 24) * Math.PI * 2) * 0.3;
-          const randomVariation = (Math.random() - 0.5) * 0.2;
-          const totalVariation = timeVariation + randomVariation;
-
-          cpu = Math.max(5, Math.min(95, Math.round(currentCPU * (1 + totalVariation))));
-
-          // Simulate scaling based on CPU if HPA exists
-          if (hpaInfo) {
-            const targetCPU = hpaInfo.targetCPUUtilization || 50;
-            if (cpu > targetCPU * 1.2) {
-              replicas = Math.min(
-                hpaInfo.maxReplicas,
-                currentReplicas + Math.floor(Math.random() * 2)
-              );
-            } else if (cpu < targetCPU * 0.7) {
-              replicas = Math.max(
-                hpaInfo.minReplicas,
-                currentReplicas - Math.floor(Math.random() * 2)
-              );
-            } else {
-              replicas = Math.max(
-                hpaInfo.minReplicas,
-                Math.min(
-                  hpaInfo.maxReplicas,
-                  currentReplicas + Math.floor((Math.random() - 0.5) * 2)
-                )
-              );
-            }
-          }
-        } else {
-          // No real CPU data - keep CPU at 0 and replicas stable
-          cpu = 0;
-          replicas = currentReplicas;
-        }
-      }
-
-      data.push({
-        time: timeString,
-        Replicas: replicas,
-        CPU: cpu,
-      });
-    }
-
-    return data.reverse(); // Reverse to get chronological order
-  };
-
-  const fetchHPAInfo = useCallback(
-    async (deploymentName: string) => {
-      if (!deploymentName || !project.namespaces?.[0]) return;
-
-      try {
-        // Find HPA that targets this deployment
-        K8s.ResourceClasses.HorizontalPodAutoscaler.apiList(
-          (hpaList: K8s.HorizontalPodAutoscaler[]) => {
-            const hpa = hpaList.find(
-              (hpa: K8s.HorizontalPodAutoscaler) =>
-                hpa.getNamespace() === project.namespaces[0] &&
-                hpa.spec?.scaleTargetRef?.name === deploymentName
-            );
-            console.log('hpa is ', hpa);
-            if (hpa) {
-              // Parse HPA CPU metrics from spec.metrics[] and status.currentMetrics[] arrays
-              const hpaJson = (hpa as any).jsonData;
-              const targetMetric = hpaJson?.spec?.metrics?.find(
-                (m: any) => m.type === 'Resource' && m.resource?.name === 'cpu'
-              );
-              const targetCPU = targetMetric?.resource?.target?.averageUtilization;
-
-              const currentMetric = hpaJson?.status?.currentMetrics?.find(
-                (m: any) => m.type === 'Resource' && m.resource?.name === 'cpu'
-              );
-              const currentCPU = currentMetric?.resource?.current?.averageUtilization;
-              const hpaData: HPAInfo = {
-                name: hpa.getName(),
-                namespace: hpa.getNamespace(),
-                minReplicas: hpa.spec?.minReplicas,
-                maxReplicas: hpa.spec?.maxReplicas,
-                targetCPUUtilization: targetCPU,
-                currentCPUUtilization: currentCPU,
-                currentReplicas: hpa.status?.currentReplicas,
-                desiredReplicas: hpa.status?.desiredReplicas,
-              };
-              setHpaInfo(hpaData);
-            } else {
-              setHpaInfo(null);
-            }
-          },
-          (error: any) => {
-            console.error('Error fetching HPA info:', error);
-            setHpaInfo(null);
-          },
-          {
-            namespace: project.namespaces[0],
-            cluster: project.clusters[0],
-          }
-        )();
-      } catch (error) {
-        console.error('Error in fetchHPAInfo:', error);
-        setHpaInfo(null);
-      }
-    },
-    [project.namespaces]
+  const { deployments, selectedDeployment, loading, error, setSelectedDeployment } = useDeployments(
+    namespace,
+    cluster
   );
+  // Find HPA that targets the selected deployment
+  const { hpaInfo } = useHPAInfo(selectedDeployment, namespace, cluster);
+  // Generate chart data based on the selected deployment and HPA info
+  const chartData = useChartData(selectedDeployment, deployments, hpaInfo);
 
-  useEffect(() => {
-    if (project.namespaces?.[0]) {
-      fetchDeployments();
-    }
-  }, [project, fetchDeployments]);
-
-  useEffect(() => {
-    if (selectedDeployment) {
-      fetchHPAInfo(selectedDeployment);
-    }
-  }, [selectedDeployment, fetchHPAInfo]);
-
-  const handleDeploymentChange = (event: any) => {
-    setSelectedDeployment(event.target.value as string);
+  const handleDeploymentChange = (deploymentName: string) => {
+    setSelectedDeployment(deploymentName);
   };
-
-  const chartData = generateChartData();
 
   return (
     <Box
@@ -285,30 +55,12 @@ function ScalingCard({ project }: ScalingCardProps) {
         }}
       >
         <Typography variant="h6">Scaling</Typography>
-        <FormControl sx={{ minWidth: 200 }} size="small" variant="outlined">
-          <InputLabel>Select Deployment</InputLabel>
-          <Select
-            value={selectedDeployment || ''}
-            onChange={handleDeploymentChange}
-            label="Select Deployment"
-            disabled={loading || deployments.length === 0}
-          >
-            {loading ? (
-              <MenuItem disabled>
-                <CircularProgress size={16} style={{ marginRight: 8 }} />
-                Loading deployments...
-              </MenuItem>
-            ) : deployments.length === 0 ? (
-              <MenuItem disabled>No deployments found</MenuItem>
-            ) : (
-              deployments.map(deployment => (
-                <MenuItem key={deployment.name} value={deployment.name}>
-                  {deployment.name}
-                </MenuItem>
-              ))
-            )}
-          </Select>
-        </FormControl>
+        <DeploymentSelector
+          selectedDeployment={selectedDeployment}
+          deployments={deployments}
+          loading={loading}
+          onDeploymentChange={handleDeploymentChange}
+        />
       </Box>
 
       {error && (
@@ -322,114 +74,14 @@ function ScalingCard({ project }: ScalingCardProps) {
       {selectedDeployment && (
         <>
           {/* Metrics Overview */}
-          <Box sx={{ mb: 2 }}>
-            <Grid container spacing={2}>
-              <Grid item xs={3}>
-                <Typography variant="body2" color="textSecondary" sx={{ fontSize: '0.75rem' }}>
-                  Scaling Mode
-                </Typography>
-                <Typography variant="h6" sx={{ fontWeight: 'bold', fontSize: '1rem' }}>
-                  {hpaInfo ? 'HPA' : 'Manual'}
-                </Typography>
-              </Grid>
-              <Grid item xs={3}>
-                <Typography variant="body2" color="textSecondary" sx={{ fontSize: '0.75rem' }}>
-                  Replica Count
-                </Typography>
-                <Typography variant="h6" sx={{ fontWeight: 'bold', fontSize: '1rem' }}>
-                  {hpaInfo?.currentReplicas ??
-                    deployments.find(d => d.name === selectedDeployment)?.readyReplicas ??
-                    'N/A'}
-                </Typography>
-              </Grid>
-              <Grid item xs={3}>
-                <Typography variant="body2" color="textSecondary" sx={{ fontSize: '0.75rem' }}>
-                  Replica Bounds
-                </Typography>
-                <Typography variant="h6" sx={{ fontWeight: 'bold', fontSize: '1rem' }}>
-                  {hpaInfo?.minReplicas !== undefined && hpaInfo?.maxReplicas !== undefined
-                    ? `${hpaInfo.minReplicas}-${hpaInfo.maxReplicas}`
-                    : 'N/A'}
-                </Typography>
-              </Grid>
-              <Grid item xs={3}>
-                <Typography variant="body2" color="textSecondary" sx={{ fontSize: '0.75rem' }}>
-                  CPU Usage
-                </Typography>
-                <Typography variant="h6" sx={{ fontWeight: 'bold', fontSize: '1rem' }}>
-                  {hpaInfo?.currentCPUUtilization !== null &&
-                  hpaInfo?.currentCPUUtilization !== undefined
-                    ? `${hpaInfo.currentCPUUtilization}%`
-                    : 'N/A'}
-                </Typography>
-              </Grid>
-            </Grid>
-          </Box>
-
+          <ScalingMetrics
+            hpaInfo={hpaInfo}
+            selectedDeployment={selectedDeployment}
+            deployments={deployments}
+          />
           {/* Chart */}
           <Box sx={{ height: 400, width: '100%' }}>
-            {chartData.length > 0 ? (
-              <ResponsiveContainer width="100%" height="100%">
-                <LineChart
-                  data={chartData}
-                  margin={{
-                    top: 10,
-                    right: 30,
-                    left: 20,
-                    bottom: 30,
-                  }}
-                >
-                  <CartesianGrid strokeDasharray="3 3" stroke="#e8e8e8" />
-                  <XAxis
-                    dataKey="time"
-                    stroke="#888"
-                    fontSize={10}
-                    tick={{ fontSize: 10 }}
-                    tickLine={{ stroke: '#e0e0e0' }}
-                    interval={5} // Show every 6th tick to avoid crowding
-                  />
-                  <YAxis
-                    stroke="#888"
-                    fontSize={10}
-                    tick={{ fontSize: 10 }}
-                    tickLine={{ stroke: '#e0e0e0' }}
-                    domain={[0, 'dataMax + 1']}
-                  />
-                  <Tooltip
-                    contentStyle={{
-                      backgroundColor: '#fff',
-                      border: '1px solid #ddd',
-                      borderRadius: '6px',
-                      fontSize: '11px',
-                      boxShadow: '0 2px 8px rgba(0,0,0,0.1)',
-                    }}
-                  />
-                  <Legend wrapperStyle={{ fontSize: '12px', paddingTop: '5px' }} />
-                  <Line
-                    type="monotone"
-                    dataKey="Replicas"
-                    stroke="#66BB6A"
-                    strokeWidth={2}
-                    dot={{ fill: '#66BB6A', strokeWidth: 0, r: 2 }}
-                    activeDot={{ r: 4, stroke: '#66BB6A', strokeWidth: 2, fill: '#fff' }}
-                  />
-                  <Line
-                    type="monotone"
-                    dataKey="CPU"
-                    stroke="#42A5F5"
-                    strokeWidth={2}
-                    dot={{ fill: '#42A5F5', strokeWidth: 0, r: 2 }}
-                    activeDot={{ r: 4, stroke: '#42A5F5', strokeWidth: 2, fill: '#fff' }}
-                  />
-                </LineChart>
-              </ResponsiveContainer>
-            ) : (
-              <Box display="flex" alignItems="center" justifyContent="center" height="100%">
-                <Typography color="textSecondary" variant="body2">
-                  No scaling data available
-                </Typography>
-              </Box>
-            )}
+            <ScalingChart chartData={chartData} />
           </Box>
         </>
       )}
